refactor(repo C15): extract single-file indexing to engine/repository/indexing.ts
Moved indexSingleFile and unlinkSingleFile, the per-file index/unlink helpers
used by the watcher integration. indexing.ts imports the content CRUD
primitives from ./content.js and imports handelize back from ../repository.js
since handelize has not yet been extracted.

Build: clean. Tests: 970/970 pass.
