+++ engine/repository/index.ts
@@ -15,3 +15,4 @@ export * from "./fts.js";
 export * from "./collections.js";
 export * from "./context-annotations.js";
 export * from "./content.js";
+export * from "./indexing.js";
